flatten run-tests
